requirements.txt
django
django-extensions
djangorestframework
django-cors-headers

